tests/wpt/tests/css/css-text/parsing/hyphenate-character-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing hyphenate-character with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#hyphenate-character">
<meta name="assert" content="hyphens supports the full grammar 'auto | &lt;string&gt;'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("hyphenate-character", "auto");
test_valid_value("hyphenate-character", "\"=\"");
test_valid_value("hyphenate-character", "\"/-/\"");
test_valid_value("hyphenate-character", "\"\\1400\"", "\"á€\"");
test_valid_value("hyphenate-character", "\"\"");
</script>
</body>
</html>
